requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
psycopg2
python-dotenv